+++ go.mod
@@ -3,11 +3,13 @@ module github.com/asecurityteam/runsqs/v2
 go 1.13
 
 require (
+	github.com/VividCortex/ewma v1.2.0 // indirect
 	github.com/asecurityteam/logevent v1.6.1
 	github.com/aws/aws-sdk-go v1.35.0
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/golang/mock v1.6.0
 	github.com/google/go-cmp v0.5.2 // indirect
+	github.com/google/uuid v1.3.0
 	github.com/justinas/alice v1.2.0 // indirect
 	github.com/kr/pretty v0.1.0 // indirect
 	github.com/rs/cors v1.7.0 // indirect
